Move user_config_file fixture to conftest.py
