+++ tests/create/test_main.py
@@ -1,9 +1,7 @@
 """Test cases for the console module."""
 import shutil
 from pathlib import Path
-from textwrap import dedent
 
-import pytest
 from click.testing import CliRunner
 
 from cutty import git
@@ -14,30 +12,6 @@ def test_main_succeeds(runner: CliRunner) -> None:
     """It exits with a status code of zero."""
     result = runner.invoke(create, ["--help"])
     assert result.exit_code == 0
-
-
-@pytest.fixture
-def cookiecutters_dir(tmp_path: Path) -> Path:
-    """A temporary directory for cookiecutters."""
-    return tmp_path / "cookiecutters_dir"
-
-
-@pytest.fixture
-def replay_dir(tmp_path: Path) -> Path:
-    """A temporary directory for replay data."""
-    return tmp_path / "replay_dir"
-
-
-@pytest.fixture
-def user_config_file(tmp_path: Path, cookiecutters_dir: Path, replay_dir: Path) -> Path:
-    """Configure cookiecutter to write to temporary directories."""
-    path = tmp_path / ".cookiecutterrc"
-    config = f"""\
-    cookiecutters_dir: {cookiecutters_dir}
-    replay_dir: {replay_dir}
-    """
-    path.write_text(dedent(config))
-    return path
 
 
 def test_create(
